created database.go file and moved DB function into it
